+++ lab4/src/kovalev/java/telecom/YAMLOperator.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 class YAMLTariff {
     private String tariffName;
     private Integer monthRub;
-    private Integer internetGB;
+    private Double internetGB;
     private Integer callMinutes;
     private Integer sms;
 }
